Use NumPy 32-bit int in `tinfo` doctest.
Appears that on Windows `int` means 32-bit int even when a 64-bit build
of Python was made. So instead of relying on `int` generally, which will
vary between 64-bit or 32-bit for Unix or Windows respectively, specify
the number of bits with a NumPy type. Also as there is a chance of
representation subtleties between Unix and Windows with a 64-bit int
(e.g. `L` postfix) chose a type that is least likely to run into these
issues in the doctest. In particular, switch to a 32-bit int from NumPy.
This both specifies a specific type of integer and picks one where its
max and min values will fall within what a 32-bit int can represent.